+++ FilmApps/app/src/main/java/com/example/filmapps/data/repository/RetrofitRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package com.example.filmapps.data.repository
 
 import com.example.filmapps.data.retrofit.GetPage
-import com.example.filmapps.presentation.model.CharacterList
 import com.example.filmapps.presentation.model.CharacterListResponce
 import java.lang.Exception
 import javax.inject.Inject
@@ -15,7 +14,7 @@ class RetrofitRepositoryImpl @Inject constructor(
     override suspend fun loadList(): CharacterListResponce {
         page++
         if(page == lastPage + 1)
-            return  CharacterListResponce.Finaly
+            return  CharacterListResponce.Finally
         return try {
             CharacterListResponce.Success(retrofit.getPage(page))
         }
