+++ FilmApps/app/src/main/java/com/example/filmapps/di/components/DetailsComponent.kt
@@ -1,0 +1,9 @@
+//package com.example.filmapps.di.components
+//
+//import com.example.filmapps.di.module.DetailsModule
+//import dagger.Component
+//
+//
+//@Component(modules = [DetailsModule::class])
+//interface DetailsComponent {
+//}
