+++ FilmApps/app/src/main/java/com/example/filmapps/di/module/CharacterListNetwork.kt
@@ -1,11 +1,10 @@
-package com.example.filmapps.domain.di.module
+package com.example.filmapps.di.module
 
 import com.example.filmapps.data.retrofit.GetPage
-import com.example.filmapps.domain.di.scope.CharacterListScope
+import com.example.filmapps.di.scope.CharacterListScope
 import dagger.Module
 import dagger.Provides
 import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 
 @Module
 class CharacterListNetwork {
