+++ FilmApps/app/src/main/java/com/example/filmapps/di/module/DatabaseModule.kt
@@ -1,4 +1,4 @@
-package com.example.filmapps.domain.di.module
+package com.example.filmapps.di.module
 
 import android.content.Context
 import androidx.room.Room
