+++ FilmApps/app/src/main/java/com/example/filmapps/di/module/DetailsModule.kt
@@ -1,0 +1,8 @@
+//package com.example.filmapps.di.module
+//
+//import dagger.Module
+//
+//
+//@Module
+//class DetailsModule {
+//}
