--- FilmApps/app/src/main/java/com/example/filmapps/domain/di/components/DetailsComponent.kt
@@ -1,9 +1,0 @@
-//package com.example.filmapps.domain.di.components
-//
-//import com.example.filmapps.domain.di.module.DetailsModule
-//import dagger.Component
-//
-//
-//@Component(modules = [DetailsModule::class])
-//interface DetailsComponent {
-//}
