--- FilmApps/app/src/main/java/com/example/filmapps/domain/di/components/FilmListComponent.kt
@@ -1,13 +1,0 @@
-package com.example.filmapps.domain.di.components
-
-import com.example.filmapps.domain.di.module.FilmListModule
-import com.example.filmapps.domain.di.module.CharacterListNetwork
-import com.example.filmapps.domain.di.scope.CharacterListScope
-import com.example.filmapps.feature.regAndAuth.presentation.viewModel.ViewModelFactory
-import dagger.Component
-
-@Component(modules = [FilmListModule::class, CharacterListNetwork::class], dependencies = [AppComponent::class])
-@CharacterListScope
-interface FilmListComponent{
-    fun viewModelsFactory(): ViewModelFactory
-}
