--- FilmApps/app/src/main/java/com/example/filmapps/domain/di/module/DetailsModule.kt
@@ -1,8 +1,0 @@
-//package com.example.filmapps.domain.di.module
-//
-//import dagger.Module
-//
-//
-//@Module
-//class DetailsModule {
-//}
