--- FilmApps/app/src/main/java/com/example/filmapps/domain/di/module/FilmListModule.kt
@@ -1,34 +1,0 @@
-package com.example.filmapps.domain.di.module
-
-import androidx.lifecycle.ViewModel
-import com.example.filmapps.data.repository.RetrofitRepository
-import com.example.filmapps.data.repository.RetrofitRepositoryImpl
-import com.example.filmapps.data.retrofit.GetPage
-import com.example.filmapps.domain.di.scope.CharacterListScope
-import com.example.filmapps.domain.useCase.GetListCharacterUseCase
-import com.example.filmapps.domain.useCase.GetListCharacterUseCaseImpl
-import com.example.filmapps.feature.regAndAuth.domain.di.module.ViewModelKey
-import com.example.filmapps.presentation.viewModel.ListCharacterViewModel
-import dagger.Binds
-import dagger.Module
-import dagger.Provides
-import dagger.multibindings.IntoMap
-import retrofit2.Retrofit
-
-
-@Module
-internal interface FilmListModule {
-
-    @Binds
-    @CharacterListScope
-    fun provideRetrofitRepositories(retrofitRepository: RetrofitRepositoryImpl) : RetrofitRepository
-
-    @Binds
-    @CharacterListScope
-    fun provideGetListCharacterUseCase(getListCharacterUseCase: GetListCharacterUseCaseImpl) : GetListCharacterUseCase
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(ListCharacterViewModel::class)
-    fun bindViewModel(listCharacterViewModel: ListCharacterViewModel): ViewModel
-}
