+++ FilmApps/app/src/main/java/com/example/filmapps/presentation/model/Character.kt
@@ -1,0 +1,3 @@
+package com.example.filmapps.presentation.model
+
+data class Character(val id: Int?, val img: String?)
