+++ FilmApps/app/src/main/java/com/example/filmapps/presentation/model/CharacterList.kt
@@ -1,8 +1,5 @@
 package com.example.filmapps.presentation.model
 
-import com.example.filmapps.data.model.DataApi
-import com.example.filmapps.presentation.viewModel.Character
-import java.lang.Exception
 
 sealed class CharacterList {
     class Success(val value: MutableList<Character>) : CharacterList()
