+++ FilmApps/app/src/main/java/com/example/filmapps/presentation/model/CharacterListResponce.kt
@@ -7,5 +7,5 @@ sealed class CharacterListResponce {
     class Success(val value: DataApi?) : CharacterListResponce()
     class Error(val message: Exception) : CharacterListResponce()
     object Loading : CharacterListResponce()
-    object Finaly : CharacterListResponce()
+    object Finally : CharacterListResponce()
 }
