+++ FilmApps/app/src/main/res/layout/fragment_character_list.xml
@@ -14,8 +14,8 @@
             android:id="@+id/image"
             android:scaleType="centerCrop"
             tools:background="@tools:sample/backgrounds/scenic"
-            android:layout_width="90dp"
-            android:layout_height="90dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
             cardView:layout_constraintStart_toStartOf="parent"
             cardView:layout_constraintTop_toTopOf="parent"/>
     </androidx.constraintlayout.widget.ConstraintLayout>
